+++ pkg/networkserver/mac/dl_channel.go
@@ -37,33 +37,43 @@ var (
 		"dl_channel", "downlink Rx1 channel frequency modification",
 		events.WithDataType(&ttnpb.MACCommand_DLChannelAns{}),
 	)()
+
+	containsDlChannel = containsMACCommandIdentifier(ttnpb.MACCommandIdentifier_CID_DL_CHANNEL)
+	consumeDlChannel  = consumeMACCommandIdentifier(ttnpb.MACCommandIdentifier_CID_DL_CHANNEL)
 )
 
 func DeviceNeedsDLChannelReqAtIndex(dev *ttnpb.EndDevice, i int) bool {
-	if i >= len(dev.MacState.DesiredParameters.Channels) {
+	currentParameters, desiredParameters := dev.MacState.CurrentParameters, dev.MacState.DesiredParameters
+	if i >= len(desiredParameters.Channels) {
 		return false
 	}
-	desiredCh := dev.MacState.DesiredParameters.Channels[i]
-	if desiredCh == nil || desiredCh.UplinkFrequency == 0 || deviceRejectedFrequency(dev, desiredCh.DownlinkFrequency) {
+	desiredCh := desiredParameters.Channels[i]
+	if desiredCh == nil ||
+		desiredCh.UplinkFrequency == 0 ||
+		deviceRejectedFrequency(dev, desiredCh.DownlinkFrequency) {
 		return false
 	}
 	if DeviceNeedsNewChannelReqAtIndex(dev, i) {
 		return desiredCh.DownlinkFrequency != desiredCh.UplinkFrequency
 	}
 	// NOTE: NewChannelReq may be needed, but parameters could have been rejected before.
-	if i >= len(dev.MacState.CurrentParameters.Channels) || dev.MacState.CurrentParameters.Channels[i] == nil {
+	if i >= len(currentParameters.Channels) || currentParameters.Channels[i] == nil {
 		return false
 	}
-	return desiredCh.DownlinkFrequency != dev.MacState.CurrentParameters.Channels[i].DownlinkFrequency
+	return desiredCh.DownlinkFrequency != currentParameters.Channels[i].DownlinkFrequency
 }
 
 func DeviceNeedsDLChannelReq(dev *ttnpb.EndDevice) bool {
 	if dev.GetMulticast() ||
-		dev.GetMacState() == nil ||
-		!macspec.UseDLChannelReq(dev.MacState.LorawanVersion) {
+		dev.GetMacState() == nil {
 		return false
 	}
-	for i := range dev.MacState.DesiredParameters.Channels {
+	macState := dev.MacState
+	if !macspec.UseDLChannelReq(macState.LorawanVersion) ||
+		containsDlChannel(macState.RecentMacCommandIdentifiers...) { // See STICKY.md.
+		return false
+	}
+	for i := range macState.DesiredParameters.Channels {
 		if DeviceNeedsDLChannelReqAtIndex(dev, i) {
 			return true
 		}
@@ -119,38 +129,50 @@ func HandleDLChannelAns(ctx context.Context, dev *ttnpb.EndDevice, pld *ttnpb.MA
 		log.FromContext(ctx).Warn("Network Server attempted to configure downlink frequency for a channel, for which uplink frequency is not defined or device is malfunctioning.")
 	}
 
-	var err error
-	dev.MacState.PendingRequests, err = handleMACResponse(ttnpb.MACCommandIdentifier_CID_DL_CHANNEL, func(cmd *ttnpb.MACCommand) error {
-		req := cmd.GetDlChannelReq()
-		if !pld.FrequencyAck {
-			if i := searchUint64(req.Frequency, dev.MacState.RejectedFrequencies...); i == len(dev.MacState.RejectedFrequencies) || dev.MacState.RejectedFrequencies[i] != req.Frequency {
-				dev.MacState.RejectedFrequencies = append(dev.MacState.RejectedFrequencies, 0)
-				copy(dev.MacState.RejectedFrequencies[i+1:], dev.MacState.RejectedFrequencies[i:])
-				dev.MacState.RejectedFrequencies[i] = req.Frequency
-			}
-		}
-		if !pld.FrequencyAck || !pld.ChannelIndexAck {
-			return nil
-		}
+	var allowMissing bool // See STICKY.md
+	dev.MacState.RecentMacCommandIdentifiers, allowMissing = consumeDlChannel(
+		dev.MacState.RecentMacCommandIdentifiers...,
+	)
 
-		if uint(req.ChannelIndex) >= uint(len(dev.MacState.CurrentParameters.Channels)) {
-			return internal.ErrCorruptedMACState.
-				WithAttributes(
-					"request_channel_id", req.ChannelIndex,
-					"channels_len", len(dev.MacState.CurrentParameters.Channels),
-				).
-				WithCause(internal.ErrUnknownChannel)
-		}
-		if dev.MacState.CurrentParameters.Channels[req.ChannelIndex] == nil {
-			return internal.ErrCorruptedMACState.
-				WithAttributes(
-					"request_channel_id", req.ChannelIndex,
-				).
-				WithCause(internal.ErrUnknownChannel)
-		}
-		dev.MacState.CurrentParameters.Channels[req.ChannelIndex].DownlinkFrequency = req.Frequency
-		return nil
-	}, dev.MacState.PendingRequests...)
+	var err error
+	dev.MacState.PendingRequests, err = handleMACResponse(
+		ttnpb.MACCommandIdentifier_CID_DL_CHANNEL,
+		allowMissing,
+		func(cmd *ttnpb.MACCommand) error {
+			req := cmd.GetDlChannelReq()
+			if !pld.FrequencyAck {
+				i := searchUint64(req.Frequency, dev.MacState.RejectedFrequencies...)
+				if i == len(dev.MacState.RejectedFrequencies) ||
+					dev.MacState.RejectedFrequencies[i] != req.Frequency {
+					dev.MacState.RejectedFrequencies = append(dev.MacState.RejectedFrequencies, 0)
+					copy(dev.MacState.RejectedFrequencies[i+1:], dev.MacState.RejectedFrequencies[i:])
+					dev.MacState.RejectedFrequencies[i] = req.Frequency
+				}
+			}
+			if !pld.FrequencyAck || !pld.ChannelIndexAck {
+				return nil
+			}
+
+			if uint(req.ChannelIndex) >= uint(len(dev.MacState.CurrentParameters.Channels)) {
+				return internal.ErrCorruptedMACState.
+					WithAttributes(
+						"request_channel_id", req.ChannelIndex,
+						"channels_len", len(dev.MacState.CurrentParameters.Channels),
+					).
+					WithCause(internal.ErrUnknownChannel)
+			}
+			if dev.MacState.CurrentParameters.Channels[req.ChannelIndex] == nil {
+				return internal.ErrCorruptedMACState.
+					WithAttributes(
+						"request_channel_id", req.ChannelIndex,
+					).
+					WithCause(internal.ErrUnknownChannel)
+			}
+			dev.MacState.CurrentParameters.Channels[req.ChannelIndex].DownlinkFrequency = req.Frequency
+			return nil
+		},
+		dev.MacState.PendingRequests...,
+	)
 	ev := EvtReceiveDLChannelAccept
 	if !pld.ChannelIndexAck || !pld.FrequencyAck {
 		ev = EvtReceiveDLChannelReject