+++ pkg/networkserver/mac/rx_param_setup.go
@@ -35,14 +35,23 @@ var (
 		"rx_param_setup", "Rx parameter setup",
 		events.WithDataType(&ttnpb.MACCommand_RxParamSetupAns{}),
 	)()
+
+	containsRxParamSetup = containsMACCommandIdentifier(ttnpb.MACCommandIdentifier_CID_RX_PARAM_SETUP)
+	consumeRxParamSetup  = consumeMACCommandIdentifier(ttnpb.MACCommandIdentifier_CID_RX_PARAM_SETUP)
 )
 
 func DeviceNeedsRxParamSetupReq(dev *ttnpb.EndDevice) bool {
-	return !dev.GetMulticast() &&
-		dev.GetMacState() != nil &&
-		(dev.MacState.DesiredParameters.Rx1DataRateOffset != dev.MacState.CurrentParameters.Rx1DataRateOffset ||
-			dev.MacState.DesiredParameters.Rx2DataRateIndex != dev.MacState.CurrentParameters.Rx2DataRateIndex ||
-			dev.MacState.DesiredParameters.Rx2Frequency != dev.MacState.CurrentParameters.Rx2Frequency)
+	if dev.GetMulticast() || dev.GetMacState() == nil {
+		return false
+	}
+	macState := dev.MacState
+	if containsRxParamSetup(macState.RecentMacCommandIdentifiers...) { // See STICKY.md.
+		return false
+	}
+	currentParameters, desiredParameters := macState.CurrentParameters, macState.DesiredParameters
+	return desiredParameters.Rx1DataRateOffset != currentParameters.Rx1DataRateOffset ||
+		desiredParameters.Rx2DataRateIndex != currentParameters.Rx2DataRateIndex ||
+		desiredParameters.Rx2Frequency != currentParameters.Rx2Frequency
 }
 
 func EnqueueRxParamSetupReq(ctx context.Context, dev *ttnpb.EndDevice, maxDownLen, maxUpLen uint16) EnqueueState {
@@ -86,19 +95,29 @@ func HandleRxParamSetupAns(ctx context.Context, dev *ttnpb.EndDevice, pld *ttnpb
 		return nil, ErrNoPayload.New()
 	}
 
+	var allowMissing bool // See STICKY.md.
+	dev.MacState.RecentMacCommandIdentifiers, allowMissing = consumeRxParamSetup(
+		dev.MacState.RecentMacCommandIdentifiers...,
+	)
+
 	var err error
-	dev.MacState.PendingRequests, err = handleMACResponse(ttnpb.MACCommandIdentifier_CID_RX_PARAM_SETUP, func(cmd *ttnpb.MACCommand) error {
-		if !pld.Rx1DataRateOffsetAck || !pld.Rx2DataRateIndexAck || !pld.Rx2FrequencyAck {
+	dev.MacState.PendingRequests, err = handleMACResponse(
+		ttnpb.MACCommandIdentifier_CID_RX_PARAM_SETUP,
+		allowMissing,
+		func(cmd *ttnpb.MACCommand) error {
+			if !pld.Rx1DataRateOffsetAck || !pld.Rx2DataRateIndexAck || !pld.Rx2FrequencyAck {
+				return nil
+			}
+
+			req := cmd.GetRxParamSetupReq()
+
+			dev.MacState.CurrentParameters.Rx1DataRateOffset = req.Rx1DataRateOffset
+			dev.MacState.CurrentParameters.Rx2DataRateIndex = req.Rx2DataRateIndex
+			dev.MacState.CurrentParameters.Rx2Frequency = req.Rx2Frequency
 			return nil
-		}
-
-		req := cmd.GetRxParamSetupReq()
-
-		dev.MacState.CurrentParameters.Rx1DataRateOffset = req.Rx1DataRateOffset
-		dev.MacState.CurrentParameters.Rx2DataRateIndex = req.Rx2DataRateIndex
-		dev.MacState.CurrentParameters.Rx2Frequency = req.Rx2Frequency
-		return nil
-	}, dev.MacState.PendingRequests...)
+		},
+		dev.MacState.PendingRequests...,
+	)
 	ev := EvtReceiveRxParamSetupAccept
 	if !pld.Rx1DataRateOffsetAck || !pld.Rx2DataRateIndexAck || !pld.Rx2FrequencyAck {
 		ev = EvtReceiveRxParamSetupReject