+++ pkg/networkserver/mac/rx_timing_setup.go
@@ -30,12 +30,21 @@ var (
 	EvtReceiveRxTimingSetupAnswer = defineReceiveMACAnswerEvent(
 		"rx_timing_setup", "Rx timing setup",
 	)()
+
+	containsRxTimingSetup = containsMACCommandIdentifier(ttnpb.MACCommandIdentifier_CID_RX_TIMING_SETUP)
+	consumeRxTimingSetup  = consumeMACCommandIdentifier(ttnpb.MACCommandIdentifier_CID_RX_TIMING_SETUP)
 )
 
 func DeviceNeedsRxTimingSetupReq(dev *ttnpb.EndDevice) bool {
-	return !dev.GetMulticast() &&
-		dev.GetMacState() != nil &&
-		dev.MacState.DesiredParameters.Rx1Delay != dev.MacState.CurrentParameters.Rx1Delay
+	if dev.GetMulticast() || dev.GetMacState() == nil {
+		return false
+	}
+	macState := dev.MacState
+	if containsRxTimingSetup(macState.RecentMacCommandIdentifiers...) { // See STICKY.md.
+		return false
+	}
+	currentParameters, desiredParameters := macState.CurrentParameters, macState.DesiredParameters
+	return desiredParameters.Rx1Delay != currentParameters.Rx1Delay
 }
 
 func EnqueueRxTimingSetupReq(ctx context.Context, dev *ttnpb.EndDevice, maxDownLen, maxUpLen uint16) EnqueueState {
@@ -71,13 +80,22 @@ func EnqueueRxTimingSetupReq(ctx context.Context, dev *ttnpb.EndDevice, maxDownL
 }
 
 func HandleRxTimingSetupAns(ctx context.Context, dev *ttnpb.EndDevice) (events.Builders, error) {
+	var allowMissing bool // See STICKY.md.
+	dev.MacState.RecentMacCommandIdentifiers, allowMissing = consumeRxTimingSetup(
+		dev.MacState.RecentMacCommandIdentifiers...,
+	)
 	var err error
-	dev.MacState.PendingRequests, err = handleMACResponse(ttnpb.MACCommandIdentifier_CID_RX_TIMING_SETUP, func(cmd *ttnpb.MACCommand) error {
-		req := cmd.GetRxTimingSetupReq()
+	dev.MacState.PendingRequests, err = handleMACResponse(
+		ttnpb.MACCommandIdentifier_CID_RX_TIMING_SETUP,
+		allowMissing,
+		func(cmd *ttnpb.MACCommand) error {
+			req := cmd.GetRxTimingSetupReq()
 
-		dev.MacState.CurrentParameters.Rx1Delay = req.Delay
-		return nil
-	}, dev.MacState.PendingRequests...)
+			dev.MacState.CurrentParameters.Rx1Delay = req.Delay
+			return nil
+		},
+		dev.MacState.PendingRequests...,
+	)
 	return events.Builders{
 		EvtReceiveRxTimingSetupAnswer,
 	}, err